requirements.txt
summa
faster-whisper
# Production dependencies for Render (pin exact versions for repeatable builds)
Flask==3.0.3
gunicorn==21.2.0
python-dotenv==1.0.1